@@ -1,8 +1,8 @@
-import { switchMap } from 'rxjs/operators';
+import { switchMap, delay, flatMap } from 'rxjs/operators';
 import { Component, OnInit, OnDestroy, Input, Output, EventEmitter } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { SubscriptionLike, Subject, of } from 'rxjs';
-import { ApiService, NodeApiRequestOptions } from '../../../../../services/api.service';
+import { ApiService } from '../../../../../services/api.service';
 import { MatDialog } from '@angular/material/dialog';
 import { SeedWordDialogComponent, WordAskedReasons } from '../../../../layout/seed-word-dialog/seed-word-dialog.component';
 import { MsgBarService } from '../../../../../services/msg-bar.service';
@@ -40,6 +40,8 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
   assistedSeedConfirmed = false;
   lastAssistedSeed = '';
   numberOfAutogeneratedWords = 0;
+  checkingCustomSeed = false;
+  checkingAssistedSeed = false;
 
   private seed: Subject<string> = new Subject<string>();
   private statusSubscription: SubscriptionLike;
@@ -68,7 +70,7 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
   }
 
   get isValid(): boolean {
-    return this.form.valid &&
+    return !this.checkingCustomSeed && this.form.valid &&
       (
         (!this.enterSeedWithAssistance && (this.customSeedIsNormal || this.customSeedAccepted)) ||
         (this.create && this.enterSeedWithAssistance && this.assistedSeedConfirmed) ||
@@ -166,8 +168,17 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
             this.partialSeed.forEach(currentWord => enteredSeed += currentWord + ' ');
             enteredSeed = enteredSeed.substr(0, enteredSeed.length - 1);
 
-            this.apiService.post('wallet/seed/verify', {seed: enteredSeed}, {useV2: true})
-              .subscribe(() => this.lastAssistedSeed = enteredSeed, () => this.msgBarService.showError('wallet.new.seed.invalid-seed-error'));
+            this.checkingAssistedSeed = true;
+            this.apiService.post('wallet/seed/verify', {seed: enteredSeed}, {useV2: true}).subscribe(
+              () => {
+                this.checkingAssistedSeed = false;
+                this.lastAssistedSeed = enteredSeed;
+              },
+              () => {
+                this.checkingAssistedSeed = false;
+                this.msgBarService.showError('wallet.new.seed.invalid-seed-error');
+              },
+            );
           }
         }
       }
@@ -261,19 +272,27 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
     }
 
     this.seedValiditySubscription = this.seed.asObservable().pipe(switchMap(seed => {
-      if ((!this.seedMatchValidator() || !this.create) && !this.enterSeedWithAssistance) {
-        return this.apiService.post('wallet/seed/verify', {seed}, {useV2: true});
+      // Verify the seed if it was entered manually and was confirmed.
+      if ((!this.seedMatchValidator() || !this.create) && !this.enterSeedWithAssistance && seed.length > 0) {
+        this.checkingCustomSeed = true;
+
+        return of(0).pipe(delay(500), flatMap(() => this.apiService.post('wallet/seed/verify', {seed}, {useV2: true})));
       } else {
         return of(0);
       }
     })).subscribe(() => {
+      this.checkingCustomSeed = false;
       this.customSeedIsNormal = true;
     }, (error: OperationError) => {
+      this.checkingCustomSeed = false;
+      // If the node said the seed is not standard, ask the user for confirmation before
+      // allowing to use it.
       error = processServiceError(error);
       if (error && error.originalError && error.originalError.status === 422) {
         this.customSeedIsNormal = false;
       } else {
         this.customSeedIsNormal = true;
+        this.msgBarService.showWarning('wallet.new.seed-checking-error');
       }
       this.subscribeToSeedValidation();
     });
