@@ -1,5 +1,5 @@
 import { Component, Input, HostListener } from '@angular/core';
-import { MatDialogRef } from '@angular/material/dialog';
+import { MatDialogRef, MatDialog } from '@angular/material/dialog';
 
 /**
  * Parent component for the content of all modal windows. It provides the title, scroll
@@ -29,9 +29,16 @@ export class ModalComponent {
     this.closePopup();
   }
 
+  constructor(
+    private matDialog: MatDialog,
+  ) { }
+
   closePopup() {
     if (!this.disableDismiss) {
-      this.dialogInternal.close();
+      // Continue only if the current modal window is the topmost one.
+      if (this.matDialog.openDialogs[this.matDialog.openDialogs.length - 1].id === this.dialogInternal.id) {
+        this.dialogInternal.close();
+      }
     }
   }
 }
