@@ -136,7 +136,7 @@ export class FormDestinationComponent implements OnInit, OnDestroy {
 
   ngOnInit() {
     this.form = this.formBuilder.group({
-      address: ['', this.showSimpleForm ? Validators.required : null],
+      address: ['', this.validateAddress.bind(this)],
       destinations: this.formBuilder.array([], this.validateDestinations.bind(this)),
     });
     this.addDestination();
@@ -362,10 +362,12 @@ export class FormDestinationComponent implements OnInit, OnDestroy {
       }
 
       // Update the hour values.
-      stringValue = dest.get('hours').value;
-      value = this.getAmount(stringValue, false);
-      if (value) {
-        this.totalHours = this.totalHours.plus(value);
+      if (this.showHourFields) {
+        stringValue = dest.get('hours').value;
+        value = this.getAmount(stringValue, false);
+        if (value) {
+          this.totalHours = this.totalHours.plus(value);
+        }
       }
     });
 
@@ -501,6 +503,18 @@ export class FormDestinationComponent implements OnInit, OnDestroy {
     });
   }
 
+  // Validates the address field.
+  private validateAddress() {
+    if (this.showSimpleForm && this.form) {
+      const address = this.form.get('address').value as string;
+      if (!address || address.trim().length < 20) {
+        return { Invalid: true };
+      }
+    }
+
+    return null;
+  }
+
   // Validates the values on the fields of the form destinations array.
   private validateDestinations() {
     if (!this.form) {
@@ -509,6 +523,13 @@ export class FormDestinationComponent implements OnInit, OnDestroy {
 
     // Check if there are invalid values.
     const invalidInput = this.destControls.find(control => {
+      if (!this.showSimpleForm) {
+        const address = control.get('address').value as string;
+        if (!address || address.trim().length < 20) {
+          return true;
+        }
+      }
+
       const controlsToCheck = ['coins'];
       if (this.showHourFields) {
         controlsToCheck.push('hours');
